test: move issue 6534 format checks to integration
